Added support for "GenericArrayType" fields that have a component type of "VariableType".
For example:
public class A<T> {
  T[] variableTypeArray;
}

With this change, Gson can serialize an instance of class "A".

We still need to provide support for the following class:
public class B<T> {
  List<T> listOfVariableTypes;
  List<T>[] arrayOfListOfVariableTypes;
}

I have submitted some tests that are currently failing, but should pass once we provide support for serializing an instance of class "B".