Added a new public method TypeAdapter.name()
This new method allows a type adapter writer to customize the JSON name.
Previously, we were using String.valueOf() to translate a key to name.
This caused a problem for using enums as keys (If a SerializedName is set, then it is ignored and toString() value is used instead).
This fixes https://github.com/google/gson/issues/1011