Test demonstrating bug in issue #1028
Gson.getDelegateAdapter() returns the wrong TypeAdapter when the
JsonAdapter annotation is used with a TypeAdapterFactory. This
happens because the appropriate TypeAdapterFactory is not already
registered, so the skipPast argument is changed to
JsonAdapterAnnotationTypeAdapterFactory. Then due to the order of the
factories, the correct one is never found.  The correct type in this
case should be CollectionTypeAdapterFactory$Adapter.

Created two tests to demonstrate this both with and without the use
of the JsonAdapter annotation.  The annotation test is marked ignore
since this bug causes it to fail.