Fix on creating TypeAdapter for Enum having abstract methods.
For enums having abstract methods, the Cass.getClass() will not return
the Class object corresponding to the enum constant. Instead of it will
return a runtime Enum class of the enum variable. This fails when
creating custom TypeAdapter for that enum. So the Type of the enum can
be obtained by Enum.getDeclaringClass() call.