Allow hierarchical Interfaces to be serialized
If the Shape class from the example above the code would implement an interface named TwoDimensionalObject with the fields x and y and some Shape should be serialized in a class that only knows that its content is a TwoDimensionalObject, this change allows that use case to work. Also, any hierarchical interfaces will work with this improvement.