Remove Test Smell: Magic Number Test
Remediation of Magic Number Test in "testSize": A Magic Number Test occurs when assert statements in a test method contain numeric literals (i.e., magic numbers) as parameters. Magic numbers do not indicate the meaning/purpose of the number. Hence, they should be replaced with constants or variables, thereby providing a descriptive name for the input.