Support for JSOG
JSOG is a very simple convention that allows arbitrary object graphs to be represented in JSON without redundancy.

During serialization, the first time an instance appears, it adds `"@id": "someId"` to the Json representation.
If it appears again, instead of re-serializing it, we can use `{"@ref": "someId"}` instead.

This removes redundancy in the JSON when the same data appears multiple times, and no duplicated
instances after deserialization.

It also allows for complex data structures with circular references to be serialized.

On the implementation details:

A Gson is now linked to a JsogPolicy, an object that decided in which objects the JSOG
protocol should be used and how to assign IDs to them.

It also adds JsogRegistry, an helper class that stores instances and their JSOG ids,
which should be used by various TypeAdapters that support JSOG.

Finally, ReflectiveTypeAdapterFactory has been modified to support JSOG.