Fix Gson.getDelegateAdapter not working correctly for @JsonAdapter field
Previously TypeAdapterFactory classes specified using the @JsonAdapter
annotation for fields were only be able to get type adapters created by the
enum and reflection-based factories when calling Gson.getDelegateAdapter(...).
This commit changes it so they are able to get type adapters created by all
factories registered on the Gson instance.

The behavior for @JsonAdapter on types (in contrast to fields) remains
unchanged.