Fix JsonReader.doPeek() advancing for malformed JSON
Previously JsonReader.doPeek() would have already advanced or modified the
state of the reader when malformed JSON was encountered and an exception
was thrown. Therefore repeatedly calling methods which peek would result
in inconsistent exceptions and could even make the JSON "valid" again by
simply advancing past the malformed JSON.

Now the state of the reader is only modified when valid JSON is encountered.
Note that there are a few remaining cases where this issue still occurs,
which will be fixed in subsequent commits.

This change required adding more JsconScope constants because if there is
enough space (or a long comment) between tokens the skipped content could
not have been stored in the buffer, e.g. `{"a":   ...     :1}`.
Here a value is expected after the `:`. However there are a lot of
whitespaces followed by a second `:` (i.e. malformed JSON).
With the previous JsonScopes it would not have been possible to represent
this state correctly. Therefore a subsequent peek would have consumed the
second `:` making the JSON "valid" again.