Make JsonReader peek mismatch exceptions use correct location
Previously when JsonReader threw an exception describing an issue with
the peeked value (e.g. it being different than the expected one) it used
the current `lineNumber` and `pos`. This resulted in confusing error
messages because both values point behind the value responsible for the
error.
For example:
  new JsonReader(new StringReader("true")).beginObject();
  // Expected BEGIN_OBJECT but was BOOLEAN at line 1 column 5 path $

With these changes the exception messages now correctly refer to the
start of the peeked value.