Throw exception when closing empty JsonTreeWriter
This makes it consistent with JsonWriter. This behavior makes sense from
the aspect that an empty string / no value is not a valid JSON document.
However, as pointed out by #421 there are cases where this might be
acceptable.
Additionally when closing the writer as part of resource clean-up of
exception handling the writer might not have written a value yet. However
in that case the exception is usually suppressed which should therefore
not be an issue.