Update to latest version of bnd-maven-plugin
This version has a couple of advantages:

1) It is aware of module-info.class files, which means we don't
need to manually override the "Require-Capability" header.

2) External packages used by reflection are now detected, i.e.
the "sun.misc" package is now added to the "Import-Package"
header. This is manually given optional resolution since it
is obviously not present on all JVMs.

3) It also introduces some new useful OSGi headers:
 - Bundle-DocURL (overriden with correct URL for manual)
 - Bundle-License
 - Bundle-SCM

This change also removes unnecessary mention of the felix
bundle-plugin in the pluginManagement section of the parent pom.

Signed-off-by: Mat Booth <mat.booth@redhat.com>