JsonWriter make indent accessible
Motivation for this change.
When using indents in the JsonWriter usually spam newlines for arrays.
I wrote a custom JsonWriter on top of this one which just implements a nicer Pretty Printing style.
Said Style disables newlines & indents when a Array is started, but enables them if a object is started within the array.
It uses a List<Boolean> to keep track of the blockage state.

The only thing that I couldn't override without adding a extra Method is the "indent". Which is set to final for reasons I assume,
but a getter should be present to aid people who want to improve their "Human Readable" JsonFile.

The idea is that you can seamlessly override the JsonWriter without the need of requiring extra functions to make original behavior work.

Thanks for reading and your time.