Fix `ProtoTypeAdapter` being unable to deserialize certain repeated fields
For example a `List<Long>` Protobuf field might internally have the
Protobuf-internal `LongList` interface type.
Previously Gson's ConstructorConstructor was nonetheless creating an
ArrayList for this, which is wrong but worked. Now with the changes in
ConstructorConstructor Gson will fail to create an instance, so this commit
tries to solve this properly in ProtoTypeAdapter.