Remove deprecation of internal fields.
The idea was to discourage people from using these, but they have been there for
a long time and the deprecation isn't going to have much effect. Meanwhile the
deprecation required `@SuppressWarnings`, which was annoying.

Also move `JsonElementTypeFactory.ADAPTER` back into `TypeAdapters` so there is
no further circular dependency between the two classes. I think it was harmless
but it's easy enough to avoid.