Fix collection tests for latest guava-testlib update (#2786)
With the fix in the latest guava-testlib version null support is now properly
checked. But it expects that if a collection does not permit null, then
creating it with null should fail as well, see guava-testlib's:
- MapCreationTester.testCreateWithNullValueUnsupported()
- CollectionCreationTester.testCreateWithNull_unsupported()

However, the previous implementation for the JsonArray and JsonObject test
was using methods which implicitly converted null to JsonNull.