Allow registering adapters for `JsonElement` again
This is done for backward compatibility, however registering the adapter still
has no effect because the built-in adapter as higher precedence (like it is
the case for multiple years already).

The Javadoc for the GsonBuilder methods has not been adjusted, to discourage
users from trying to register adapters for JsonElement.