Add adapters for Guava's immutable ccollections.
These adapters are added automatically, but only if Guava is on the
classpath. They correctly deserialize immutable collections and maps, by
creating a builder and adding the incoming elements to it, before
finally building the immutable object.

The principal motivation for this change is that #2068, if applied, will
trigger a large number of failures from Google's internal tests. Code is
incorrectly deserializing `TypeToken<ImmutableList<String>>` and the like, and
getting away with it because it doesn't happen to perform any actions that
would notice that the deserialized object is actually an `ArrayList`.