Fix ConstructorConstructor creating mismatching Collection and Map instances (#2068)
* Fix ConstructorConstructor creating mismatching Collection instances

* Support EnumMap deserialization

* Use LinkedTreeMap for maps with String supertype as key

* Fix test method name typo

* Fix `ProtoTypeAdapter` being unable to deserialize certain repeated fields

For example a `List<Long>` Protobuf field might internally have the
Protobuf-internal `LongList` interface type.
Previously Gson's ConstructorConstructor was nonetheless creating an
ArrayList for this, which is wrong but worked. Now with the changes in
ConstructorConstructor Gson will fail to create an instance, so this commit
tries to solve this properly in ProtoTypeAdapter.

* Adjust tests

* Don't use LinkedTreeMap for String supertypes as key & small test fix

This reverts a previous commit of this pull request, and matches the
original behavior again.

* Fix typo in exception message

* Adjust ConstructorConstructor for Java 8

* Use lambdas to make it more explicit which constructor overloads are used

* Don't use Unsafe to create Collection and Map instances