Update sources to satisfy a new Error Prone check. (#2887)
* Update sources to satisfy a new Error Prone check.

The latest Error Prone release includes a new check
`EffectivelyPrivate`, which flags class members that are public or
protected when the enclosing class isn't. This is only a warning in
Error Prone, but because we treat warnings as errors it currently breaks
the build.

Nearly all cases were in private nested classes in tests. For the
most part the members didn't need to be public. In a couple of places
they were accessed through reflection in the test, and it was enough
to change `getField` to `getDeclaredField` (etc). (`getField` only
retrieves public fields.) I don't believe any of these tests depended on
the public members for the correctness of what they were testing.
Although `getField` works when the target field is public, regardless of
the visibility of the containing class, actually reading or writing a
field requires both the field and its containing class to be public
unless `Field.setAccessible(true)` is called, and likewise for
constructors and methods.