Small modifications to make building a subset possible.
The changes here make it possible to build a subset of Gson that includes `JsonReader`, `JsonWriter`, `JsonParser`, and `JsonElement` and its subclasses, but does not include the `Gson` class or the standard `TypeAdapter` classes. The idea is that this subset avoids the problematic uses of reflection that make Gson inappropriate for use on Android and similar platforms.

This change makes the class `com.google.gson.internal.bind.JsonElementTypeAdapter` public, which is convenient but of course means that people are going to start referencing it despite the `.internal.` in its name. However, they can already reference `TypeAdapters.JSON_ELEMENT` in the same directory, so we're not really exposing anything new.