Update `OSGiManifestIT` not to depend on order.
It appears that sometimes the clauses being checked for appear in the
other order, possibly because of a `HashMap` or the like somewhere in
the guts of OSGi. I haven't seen this on GitHub, but I do see it when
running locally with Google's JDK, which has more hash randomization.