Update `test-graal-native-image`.
With the latest dependencies, this was failing with this message:
```
Error: Classes that should be initialized at run time got initialized during image building:
 org.junit.jupiter.engine.discovery.MethodSegmentResolver was unintentionally initialized at build time.
```

I don't know enough about GraalVM to be sure of what the best fix is,
but Gemini suggested the thing I'm doing here. It seems to work.